# Migrate crossKernel from THC to ATen (CUDA)