--- a/aten/src/THC/THCTensorMath.h
+++ b/aten/src/THC/THCTensorMath.h
@@ -22,12 +22,6 @@
 #include <THC/generic/THCTensorMathPairwise.h>
 #include <THC/THCGenerateBoolType.h>
 
-#include <THC/generic/THCTensorMathPointwise.h>
-#include <THC/THCGenerateAllTypes.h>
-
-#include <THC/generic/THCTensorMathPointwise.h>
-#include <THC/THCGenerateBoolType.h>
-
 #include <THC/generic/THCTensorScatterGather.h>
 #include <THC/THCGenerateAllTypes.h>