--- a/aten/src/THC/generated/THCTensorMathPointwiseByte.cu
+++ b/aten/src/THC/generated/THCTensorMathPointwiseByte.cu
@@ -1,5 +1,0 @@
-#include <THC/THCTensorMathPointwise.cuh>
-#include <THC/THCTensor.hpp>
-
-#include <THC/generic/THCTensorMathPointwise.cu>
-#include <THC/THCGenerateByteType.h>