--- a/aten/src/THC/generated/THCTensorMathPointwiseHalf.cu
+++ b/aten/src/THC/generated/THCTensorMathPointwiseHalf.cu
@@ -1,5 +1,0 @@
-#include <THC/THCTensorMathPointwise.cuh>
-#include <THC/THCTensor.hpp>
-
-#include <THC/generic/THCTensorMathPointwise.cu>
-#include <THC/THCGenerateHalfType.h>