Ref  #24507 (There doesn't seem to be an actual issue for cross)

This also moves the remaining operator functors in `THCTensorMathPointwise.cuh`  to `SparseCUDATensorMath.cu` which is the only file using them.