@peterbell10 2021-06-15
I measure a performance improvement here:
|        Shape       |  Before |  After  |
|:------------------:|:-------:|:-------:|
| (100, 100, 100, 3) | 99.2 us | 93.9 us |
|    (100, 100, 3)   | 4.91 us | 3.91 us |